Refactor: Flatten out src/ structure and remove mpy files
- mpy files removed in favor of adafruit library dependencies
- remove lib/ and mode/ directories under src/
