+++ .gitignore
@@ -1,1 +1,4 @@
+.venv/
+*.egg-info/
+
 __pycache__
