--- src/lib/modes/__init__.py
@@ -1,4 +1,0 @@
-# lib/modes/__init__.py
-
-# This file can be left empty or used to import classes for easier access
-
