+++ src/solar_simulator.py
@@ -1,39 +1,40 @@
 try:
-    from ulab import numpy as np  # Use ulab when running on Pico
+    from ulab import numpy as np
 except ImportError:
-    import numpy as np  # Use numpy when running on PC
-import adafruit_mcp4728 as MCP  # 12-bit DAC
-import adafruit_ads1x15.ads1015 as ADS  # 4-channel ADC
+    import numpy as np
+
+import adafruit_mcp4728 as MCP
+import adafruit_ads1x15.ads1015 as ADS
 from adafruit_ads1x15.analog_in import AnalogIn
 import digitalio as dio
 from busio import I2C
 from pwmio import PWMOut
 import board
+import math
 
 MAX_VALUE = 65535
 
 
 class SolarSimulator:
-    # Initializes the Solar Simulator module
     def __init__(self, pwm_freq: int = 5000, verbose: int = 0):
-        # Initialize constants
+        """Initialize a SolarSimulator object."""
+
         if verbose: print("Initializing simulator...")
+
         self.PWM_FREQ = pwm_freq
         self.verbose = verbose
         self.peak = 0.3
         self.i2c = I2C(board.GP27, board.GP26)
-        if self.verbose >= 2: print(f"I2C initialized, addresses found: {self.__portScan()}")
-        # Create ADS and MCP objects
+        if self.verbose >= 2: print(f"I2C initialized, addresses found: {self._portScan()}")
+
         self.ads = ADS.ADS1015(self.i2c)
         self.mcp = MCP.MCP4728(self.i2c)
         if verbose: print("Initialized I2C devices")
 
-        # Create the Halogen object
         self.hal = PWMOut(board.GP28, frequency=self.PWM_FREQ, duty_cycle=0, variable_frequency=True)
         if verbose: print(f"Initialized PWM at {self.PWM_FREQ}Hz")
 
         self.therm_safe = True
-        # Initialize current light settings to zero
         self.current_light_settings = {
             'v': 0,
             'w': 0,
@@ -48,8 +49,9 @@ class SolarSimulator:
 
         if verbose: print("Solar Simulator initialized")
 
-    # Sets the LEDs and halogen brightness as a 16-bit integer value
+
     def setLEDs(self, v: int = 0, w: int = 0, c: int = 0, h: int = 0):
+        """Set the LEDs and halogen brightness as a 16-bit integer value."""
         self.mcp.channel_a.value = v
         self.mcp.channel_b.value = w
         self.mcp.channel_c.value = c
@@ -62,6 +64,7 @@ class SolarSimulator:
             'c': c,
             'h': h,
         }
+
         if self.verbose >= 2:
             print(
                 f"VIOLET: {self.mcp.channel_a.value}, WHITE: {self.mcp.channel_b.value}, CYAN: {self.mcp.channel_c.value}, HAL: {self.hal.duty_cycle}")
@@ -69,41 +72,45 @@ class SolarSimulator:
             print(
                 f"VIOLET: {self.mcp.channel_a.value // 655}%, WHITE: {self.mcp.channel_b.value // 655}%, CYAN: {self.mcp.channel_c.value // 655}%, HAL: {self.hal.duty_cycle // 655}%")
 
-    #65535，percentage just do that
-    # Returns a list of thermal values per thermistor channel in Celsius
+
     def checkThermals(self) -> list:
+        """Return a list of thermal values per thermistor channel in Celsius."""
         thermals = []
-        thermistors = self.__readThermistors()
+        thermistors = self._readThermistors()
+
         for i, chan in zip(range(3), thermistors):
             thermals.append(chan[2])
             if self.verbose == 1: print(f"Channel[{i}]: {chan[2]}C")
+
         return thermals
 
 
-    # Helper function that prints all available I2C devices
-    def __portScan(self) -> list:
+    def _portScan(self) -> list:
+        """Helper function that prints all available I2C devices."""
         self.i2c.try_lock()
         found = self.i2c.scan()
         self.i2c.unlock()
         return [hex(i) for i in found]
 
-    # Helper function that reads all of the thermistors and returns a list of lists of data for each channel
-    # Output: [[CHAN0 binary data, voltage, celsius temp], [CHAN1 binary data, voltage, celsius temp], [CHAN2 binary data, voltage, celsius temp]]
-    def __readThermistors(self) -> list:
+
+    def _readThermistors(self) -> list:
+        """Helper function that reads all of the thermistors and returns a list of lists of data for each channel.
+
+        Output: [[CHAN0 binary data, voltage, celsius temp],
+                 [CHAN1 binary data, voltage, celsius temp],
+                 [CHAN2 binary data, voltage, celsius temp]]
+        """
         therm_values = []
+
         for i in range(0, 3):
             chan = AnalogIn(self.ads, i)
             therm_values.append([chan.value >> 4, chan.voltage, calcTemp(chan.voltage)])
             if self.verbose >= 2: print(f"Channel[{i}]: {chan.value >> 4}, {chan.voltage}v, {calcTemp(chan.voltage)}C")
+
         return therm_values
 
 
-
-
-import math
-
 def calcTemp(v_adc: float, vcc: float = 3.3, r_fixed: float = 10000.0) -> float | None:
-
     # Compute thermistor resistance
     r_therm = r_fixed * v_adc / (vcc - v_adc)
 
@@ -118,7 +125,6 @@ def calcTemp(v_adc: float, vcc: float = 3.3, r_fixed: float = 10000.0) -> float 
         return temp_c
     except ValueError:
         return None  # math domain error if log input is negative
-
 
 
 def calcSteps(limiter: float = 1) -> list:
