+++ src/thermistor_helper.py
@@ -1,5 +1,3 @@
-# Thermistor conversion helper
-# TODO: Compile to `.mpy`
 import ulab.numpy as np
 
 # Takes a thermistor's voltage and returns it's temperature in Celsius
